implement bot list posting

+++ scripty_botlists/src/lists/botlist_me/models.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Serialize)]
 pub struct PostStats {
-	pub server_count: u64,
-	pub shard_count:  u64,
+	pub server_count: usize,
+	pub shard_count:  u32,
 }
 
 #[derive(Debug, Deserialize, Copy, Clone)]
