+++ scripty_botlists/src/lists/infinitybots_gg/models.rs
@@ -1,5 +1,5 @@
 #[derive(Serialize, Debug, Copy, Clone)]
 pub struct PostStats {
-	pub servers: u64,
-	pub shards:  u64,
+	pub servers: usize,
+	pub shards:  u32,
 }
