+++ scripty_botlists/src/lists/radarcord_net/models.rs
@@ -1,7 +1,7 @@
 #[derive(Serialize, Debug, Copy, Clone)]
 pub struct PostStats {
-	pub servers: u64,
-	pub shards:  u64,
+	pub servers: usize,
+	pub shards:  u32,
 }
 
 #[derive(Deserialize, Debug, Clone)]
