+++ scripty_botlists/src/lists/top_gg/models.rs
@@ -16,6 +16,6 @@ pub enum VoteWebhookType {
 
 #[derive(Debug, Serialize)]
 pub struct PostStats {
-	pub server_count: u64,
-	pub shard_count:  u64,
+	pub server_count: usize,
+	pub shard_count:  u32,
 }
